+++ src/main/scala/io/viash/config/ArgumentGroup.scala
@@ -21,13 +21,78 @@ import io.viash.config.arguments.Argument
 import io.viash.schemas._
 
 @description("A grouping of the @[arguments](argument), used to display the help message.")
+@example(
+  """argument_groups:
+    |  - name: "Input"
+    |    arguments:
+    |      - name: "--id"
+    |        type: string
+    |        required: true
+    |      - name: "--input"
+    |        type: file
+    |        required: true
+    |  - name: "Output"
+    |    arguments:
+    |      - name: "--output"
+    |        type: file
+    |        direction: output
+    |        required: true
+    |      - name: "--output_optional"
+    |        type: file
+    |        direction: output
+    |""".stripMargin,
+    "yaml")
 case class ArgumentGroup(
   @description("The name of the argument group.")
   name: String,
 
-  @description("Description of foo`, a description of the argument group. Multiline descriptions are supported.")
+  @description("A clean version of the argument group's name. This is only used for documentation.")
+  @example("label: \"My argument group\"", "yaml")
+  @default("Empty")
+  @since("Viash 0.9.0")
+  label: Option[String] = None,
+
+  @description("A one-sentence summary of the argument group. This is only used for documentation.")
+  @example("summary: \"Arguments related to functionality XYZ\"", "yaml")
+  @default("Empty")
+  @since("Viash 0.9.0")
+  summary: Option[String] = None,
+
+  @description("A description of the argument group. This is only used for documentation. Multiline descriptions are supported.")
+  @example(
+    """description: |
+      |  A (multiline) description of the purpose of the arguments
+      |  in this argument group.""".stripMargin, "yaml")
+  @default("Empty")
   description: Option[String] = None,
 
-  @description("List of arguments.")
+  @description(
+    """A list of @[arguments](argument) for this component. For each argument, a type and a name must be specified. Depending on the type of argument, different properties can be set. See these reference pages per type for more information:  
+      |
+      | - @[string](arg_string)
+      | - @[file](arg_file)
+      | - @[integer](arg_integer)
+      | - @[double](arg_double)
+      | - @[boolean](arg_boolean)
+      | - @[boolean_true](arg_boolean_true)
+      | - @[boolean_false](arg_boolean_false)
+      |""".stripMargin)
+  @example(
+    """arguments:
+      |  - name: --foo
+      |    type: file
+      |    alternatives: [-f]
+      |    description: Description of foo
+      |    default: "/foo/bar"
+      |    must_exist: true
+      |    direction: output
+      |    required: false
+      |    multiple: true
+      |    multiple_sep: ";"
+      |  - name: --bar
+      |    type: string
+      |""".stripMargin,
+      "yaml")
+  @default("Empty")
   arguments: List[Argument[_]] = Nil
 )