+++ src/main/scala/io/viash/config/arguments/DoubleArgument.scala
@@ -46,7 +46,24 @@ case class DoubleArgument(
   @default("Empty")
   alternatives: OneOrMore[String] = Nil,
 
-  @description("A description of the argument. This will be displayed with `--help`.")
+  @description("A clean version of the argument's name. This is only used for documentation.")
+  @example("label: \"My argument\"", "yaml")
+  @default("Empty")
+  @since("Viash 0.9.0")
+  label: Option[String] = None,
+
+  @description("A one-sentence summary of the argument. This is only used for documentation.")
+  @example("summary: \"This argument sets XYZ.\"", "yaml")
+  @default("Empty")
+  @since("Viash 0.9.0")
+  summary: Option[String] = None,
+
+  @description("A description of the argument. This is only used for documentation. Multiline descriptions are supported.")
+  @example(
+    """description: |
+      |  A (multiline) description of the purpose of
+      |  this argument.""".stripMargin, "yaml")
+  @default("Empty")
   description: Option[String] = None,
 
   @description("Structured information. Can be any shape: a string, vector, map or even nested map.")
@@ -139,6 +156,8 @@ case class DoubleArgument(
     `type`: String, 
     name: String, 
     alternatives: OneOrMore[String],
+    label: Option[String],
+    summary: Option[String],
     description: Option[String],
     info: Json,
     example: OneOrMore[Double],
@@ -149,6 +168,6 @@ case class DoubleArgument(
     multiple_sep: String,
     dest: String
   ): Argument[Double] = {
-    copy(name, alternatives, description, info, example, default, required, this.min, this.max, direction, multiple, multiple_sep, dest, `type`)
+    copy(name, alternatives, label, summary, description, info, example, default, required, this.min, this.max, direction, multiple, multiple_sep, dest, `type`)
   }
 }